Split generic diagnostic annotation tests
