+++ docs/PHASE_PLAN.md
@@ -201,6 +201,9 @@ checked-in docs, tests, and commits only.
 - Typechecker monomorphization template dependency install/restore now lives in
   a dedicated helper module, keeping generic specialization flow separate from
   temporary source-module dependency overlays.
+- Generic diagnostic integration coverage now keeps nested/function/container
+  annotation arity tests in a focused module, separate from direct generic
+  call, method, local, and declaration annotation arity cases.
 - Module-system unit tests now live beside the module-system implementation in
   a dedicated test module, keeping module graph/load entry points compact.
 - Resolver expected-local traversal now lives in a dedicated validation support
