+++ tests/generic_diagnostics.rs
@@ -32,6 +32,8 @@ mod annotations;
 mod behavior_impls;
 #[path = "generic_diagnostics/bounds.rs"]
 mod bounds;
+#[path = "generic_diagnostics/composite_annotations.rs"]
+mod composite_annotations;
 #[path = "generic_diagnostics/inference_conflicts.rs"]
 mod inference_conflicts;
 #[path = "generic_diagnostics/method_type_args.rs"]
